requirements.txt
pydantic==2.9.2
pandas==2.2.3
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pgvector==0.3.4
tqdm==4.66.5
lxml==5.3.0
pydantic==2.9.2
SQLAlchemy==2.0.35
spacy==3.8.2
openai==1.51.0
pytest==8.3.3